# perf: refactor escapeRegExp and removeDuplicateSlashes functions